requirements.txt
numpy==2.3.0
paho-mqtt==2.1.0
python-dotenv==1.1.1
requests==2.32.5
ta==0.11.0